% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_tsne.R
\name{fftRtsne}
\alias{fftRtsne}
\title{ffRtsne}
\usage{
fftRtsne(X, dims = 2, perplexity = 30, theta = 0.5,
  check_duplicates = TRUE, max_iter = 1000, fft_not_bh = TRUE,
  ann_not_vptree = TRUE, stop_lying_iter = 250, exaggeration_factor = 12,
  no_momentum_during_exag = FALSE, start_late_exag_iter = -1,
  late_exag_coeff = 1, n_trees = 50, search_k = -1, rand_seed = -1,
  nterms = 3, intervals_per_integer = 1, min_num_intervals = 50,
  data_path = NULL, result_path = NULL, fast_tsne_path = "bin/fast_tsne",
  nthreads = 0, ...)
}
