% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnehm.R
\name{tsnehm}
\alias{tsnehm}
\title{tSNE Heatmap}
\usage{
tsnehm(expression_matrix, goi, tsne_embedding, cell_labels, enrich = 0,
  breaks = 100)
}
\arguments{
\item{expression_matrix}{Full expression matrix genes (rows) vs. cells (columns). Rownames should be gene names}

\item{goi}{goi Genes of Interest}

\item{tsne_embedding}{1D t-SNE embedding}

\item{cell_labels}{Labels for each cell. Typically these are the cluster assignments for each cell, as obtained by 
dbscan on the 2D t-SNE. This way, the columns of the heatmap will have a color assigned to each of them,
and they can be mapped to a corresponding location on the 2D t-SNE}

\item{enrich}{For every gene in goi, find enrich number of genes that are close to that gene in the distance induced by the 1D t-SNE}

\item{breaks}{Number of bins
import gplots might not be required}
}
\description{
Computes t-SNE Heatmap (EXPERIMENTAL)
}
